feat(contented): add image processing (#634)
#### What this PR does / why we need it:

Add images support for contented.

Image commonly used in Markdown is supported but with some caveats.
Contented being a prose-bundler is designed to bundle prose such that it
is portable and can be deployed to different sites or domains. This
makes it difficult to support images that are not bundled together with
the prose. To get around this, all local images are embedded into the
Markdown file as base64 encoded strings. Remote images are left as is.